Refactor C typeof machinery into a separate file _typeof.c
