+++ setup.py
@@ -62,10 +62,12 @@ ext_npymath_exports = Extension(name='numba._npymath_exports',
 ext_dispatcher = Extension(name="numba._dispatcher",
                            include_dirs=[numpy.get_include()],
                            sources=['numba/_dispatcher.c',
+                                    'numba/_typeof.c',
                                     'numba/_dispatcherimpl.cpp',
                                     'numba/typeconv/typeconv.cpp'],
                            depends=["numba/_pymodule.h",
-                                    "numba/_dispatcher.h"],
+                                    "numba/_dispatcher.h",
+                                    "numba/_typeof.h"],
                            extra_link_args=cpp_link_args)
 
 ext_helperlib = Extension(name="numba._helperlib",
